Added changes for search and responseEntity

+++ src/main/java/com/josh/asset_managment_system/Admin/AdminController.java
@@ -2,25 +2,22 @@ package com.josh.asset_managment_system.Admin;
 
 
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-
 import com.josh.asset_managment_system.Employee.*;
-
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
-
 import com.josh.asset_managment_system.Asset.*;
 import com.josh.asset_managment_system.AssetRequest.AssetRequest;
 import com.josh.asset_managment_system.AssetRequest.AssetRequestService;
@@ -44,35 +41,56 @@ public class AdminController {
     // *********************   Employee API's ******************//
     
     @PostMapping("/createEmployee")
-    public String createEmployee(@RequestParam String userName, @RequestParam String employeeName, @RequestParam String password) {
+    public ResponseEntity<?> createEmployee(@RequestBody Employee employee) {
     	
     	
-        Employee e =  employeeService.createEmployee(employeeName, userName, password);
+    	if (employee.getEmpName()== null || employee.getEmpName().isBlank()) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please send employee name correctly");
+    	}else if (employee.getPassword()== null || employee.getPassword().isBlank()) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please send password correctly");
+    	}else if (employee.getUserName()== null || employee.getUserName().isBlank()) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please send user name correctly");
+    	}
+    	
+        Employee emp =  employeeService.createEmployee(employee);
         
-        if (!e.getUserName().isEmpty()) {
+        if (!emp.getUserName().isEmpty()) {
         	EmailDetails details = new EmailDetails();
         	StringBuilder msg = new StringBuilder();
-        	msg.append("Your login credentials are Username: "+e.getUserName()+"\n Password:"+e.getPassword());
+        	msg.append("Your login credentials are \n Username: "+emp.getUserName()+"\n Password:"+emp.getPassword());
         	details.setMsgBody(msg.toString());
-        	details.setRecipient(e.getUserName());
+        	details.setRecipient(emp.getUserName());
         	details.setSubject("Your account has been created");
-        	return emailService.sendSimpleMail(details);
+        	String result = emailService.sendSimpleMail(details);
+        	return ResponseEntity.status(HttpStatus.OK).body(result);
+        	
         } else {
-        	  return "Mail not send";
+        	return ResponseEntity.status(HttpStatus.OK).body("Please try again");
         } 
     }
     
     @DeleteMapping("/deleteEmployee")
-    public String deleteEmployee(@RequestParam String userName) {
+    public ResponseEntity<?> deleteEmployee(@RequestParam String userName) {
+    	if (userName == null || userName.isBlank()) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please send user name correctly");
+    	}
     	
         employeeService.deleteEmployee(userName);
-        return "Employee deleted successfully";
+        return ResponseEntity.status(HttpStatus.OK).body("Employee deleted successfully");
          
     }
     
+    
     @GetMapping("/getEmployeeList")
-    public List<Employee> getEmployeeList() {
-        return employeeService.getAllEmployeeList();
+    public ResponseEntity<?> getEmployeeList() {
+    	
+    	List<Employee> employeeList =  employeeService.getAllEmployeeList();
+    	if(employeeList.isEmpty()) {
+    		return ResponseEntity.status(HttpStatus.NO_CONTENT).body("No employee created yet,Please create the employee");
+    		
+    	}
+       return  ResponseEntity.ok(employeeList);
+
     }
     
     
@@ -80,43 +98,112 @@ public class AdminController {
     // *********************   Asset API's ******************//
     
     @GetMapping("/getAssetList")
-    public List<Asset> getAssetList() {
-        return assetService.getAllAssets();
+    public ResponseEntity<?> getAssetList() {
+    	List<Asset> assetList = assetService.getAllAssets();
+    	if (assetList == null || assetList.isEmpty()) {
+    		return ResponseEntity.status(HttpStatus.NO_CONTENT).body("No assets found");
+    	} else {
+    		
+    		 return ResponseEntity.status(HttpStatus.OK).body(assetList);
+    	} 
     }
     
     @PostMapping("/createAsset")
-    public Asset createAsset(@RequestParam String assetName) {
-        return assetService.createAsset(assetName);
+    public ResponseEntity<?> createAsset(@RequestBody Map<String,String> assetObject) {
+    	
+    	if (assetObject.get("assetName") == null || assetObject.get("assetName").isBlank()) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please send asset Name correctly");
+    	} else {
+    		 assetService.createAsset(assetObject.get("assetName"));
+    		 
+    		 return ResponseEntity.status(HttpStatus.OK).body("Asset created successfully");
+    	} 
+    	
     }
     
-    @DeleteMapping("/deleteAsset/assetId={assetId}")
-    public String deleteAsset(@PathVariable Integer assetId) {
-        return assetService.deleteAsset(assetId);
+    
+    
+    @DeleteMapping("/deleteAsset")
+    public ResponseEntity<?> deleteAsset(@RequestParam Integer assetId) {
+    	
+    	if (assetId == null || assetId <= 0) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please enter correct asset ID");
+    	} else {
+    		assetService.deleteAsset(assetId);
+    		 return ResponseEntity.status(HttpStatus.OK).body("asset deleted successfully");
+    	} 
+    }
+    
+    @GetMapping("/searchAssetNameInAssetList/assetName={assetName}")
+    public ResponseEntity<?> searchAssetNameInAssetList(@PathVariable String assetName) {
+    	
+    	if (assetName == null || assetName.isBlank()) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please enter correct asset name");
+    	}
+    	
+    	List<Asset>  assetList =  assetService.searchAssets(assetName);
+    	
+    	if (assetList == null || assetList.isEmpty()) {
+    		return ResponseEntity.status(HttpStatus.NO_CONTENT).body("No searched assets found");
+    	} else {
+    		
+    		 return ResponseEntity.status(HttpStatus.OK).body(assetList);
+    	} 
+    	
     }
     
     
     
     @PostMapping("/allocateAsset")
-    public String updateAsset(@RequestParam Integer assetId,@RequestParam Integer empId) {
-        return assetService.allocateAsset(assetId, empId);
+    public ResponseEntity<?> updateAsset(@RequestBody Asset asset) {
+    	
+    	if (asset.getAssetId() == null || asset.getAssetId()<= 0) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please enter correct asset ID");
+    	}else if (asset.getEmpId() == null || asset.getEmpId()<= 0) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please enter correct employee ID");
+    	}
+    	 return ResponseEntity.status(HttpStatus.OK).body(assetService.allocateAsset(asset));
     }
     
     @PostMapping("/updateAssetName")
-    public String updateAssetName(@RequestParam Integer assetId,@RequestParam String assetName) {
-        return assetService.updateAsset(assetId, assetName);
+    public ResponseEntity<?> updateAssetName(@RequestBody Asset asset) {
+    	
+    	if (asset.getAssetId() == null || asset.getAssetId()<= 0) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please enter correct asset ID");
+    	}else if (asset.getAssetName() == null || asset.getAssetName().isBlank()) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please enter correct asset name");
+    	}
+    	return ResponseEntity.status(HttpStatus.OK).body(assetService.updateAsset(asset));
+      
     }
     
     
     // *********************   Asset List API's ******************//
     
     @GetMapping("/getAssetRequestList")
-    public List<AssetRequest> getAssetRequestList() {
-        return assetRequestService.getAllAssetRequests();
+    public ResponseEntity<?> getAssetRequestList() {
+    	
+    	List<AssetRequest> assetRequestList =assetRequestService.getAllAssetRequests();
+    	if (assetRequestList == null || assetRequestList.isEmpty()) {
+    		return ResponseEntity.status(HttpStatus.NO_CONTENT).body("No assets found");
+    	} else {
+    		
+    		 return ResponseEntity.status(HttpStatus.OK).body(assetRequestList);
+    	} 
+    	
     }
     
     @PostMapping("/changeAssetRequestStatus")
-    public String changeAssetRequestStatus(@RequestParam Integer requestId,@RequestParam String status) {
-        return assetRequestService.changeAssetRequestStatus(requestId,status);
+    public ResponseEntity<?> changeAssetRequestStatus(@RequestBody AssetRequest request) {
+    	
+    	if (request.getRequestId() == null || request.getRequestId()<= 0) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please enter correct asset ID");
+    	}else if (request.getAllocationStatus() == null || request.getAllocationStatus().isBlank()) {
+    		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Please enter correct status");
+    	}
+    	
+    	return ResponseEntity.status(HttpStatus.OK).body(assetRequestService.changeAssetRequestStatus(request));
+      
     }
